split components out into /components directory
