+++ packages/react/src/recipes/FlexTemplate/FlexSection/FlexSection.tsx
@@ -9,33 +9,24 @@ import {
   Box,
   BreakoutBanner,
   Card,
-  FrostedGlassVFX,
   Grid,
   Heading,
   Icon,
   Image,
   InlineLink,
-  Label,
   Link,
   LogoSuite,
   Pillar,
   PricingOptions,
   Prose,
-  River,
-  RiverAccordion,
-  RiverBreakout,
   RiverStoryScroll,
   Section,
   SectionIntro,
   Stack,
   Statistic,
-  Testimonial,
   Text,
-  Timeline,
-  useTheme,
 } from '../../..'
 
-import monaAvatar from '../../../fixtures/images/avatar-mona.png'
 import pinterestLogo from '../../../fixtures/images/logos/pinterest.png'
 import shopifyLogo from '../../../fixtures/images/logos/shopify.png'
 import twilioLogo from '../../../fixtures/images/logos/twilio.png'
@@ -44,12 +35,9 @@ import vercelLogo from '../../../fixtures/images/logos/vercel.png'
 import placeholderImage from '../../../fixtures/images/placeholder.png'
 import lightNarrowBg from '../../../fixtures/images/light-vertical-banner.png'
 import lightWideBg from '../../../fixtures/images/light-horizontal-banner.png'
-import startShapeLight from '../../../fixtures/images/testimonial-bg-1.png'
-import endShapeLight from '../../../fixtures/images/testimonial-bg-2.png'
-import startShapeDark from '../../../fixtures/images/testimonial-bg-1-dark.png'
-import endShapeDark from '../../../fixtures/images/testimonial-bg-2-dark.png'
 
 import styles from './FlexSection.module.css'
+import {FlexSectionTestimonials, ContentfulRiver, ContentfulRiverBreakout, ContentfulRiverAccordion} from './components'
 
 type FlexSectionProps = {
   component
@@ -692,384 +680,5 @@ export function FlexSection({component, className}: FlexSectionProps) {
         </Stack>
       </Box>
     </Section>
-  )
-}
-
-export const backgroundStylesMap = {
-  Productivity: {
-    className: styles.productivity,
-    marginBlockStart: 'none',
-    marginBlockEnd: 16,
-    left: {
-      width: 444,
-    },
-    right: {
-      width: 574,
-    },
-  },
-  Collaboration: {
-    className: styles.collaboration,
-    marginBlockStart: 24,
-    marginBlockEnd: 'none',
-    left: {
-      width: 426,
-    },
-    right: {
-      width: 417,
-    },
-  },
-  AI: {
-    className: styles.ai,
-    marginBlockStart: 'none',
-    marginBlockEnd: 'none',
-    left: {
-      width: 424,
-    },
-    right: {
-      width: 433,
-    },
-  },
-  Security: {
-    className: styles.security,
-    marginBlockStart: 8,
-    marginBlockEnd: 'none',
-    left: {
-      width: 349,
-    },
-    right: {
-      width: 521,
-    },
-  },
-  Enterprise: {
-    className: styles.enterprise,
-    marginBlockStart: 'none',
-    marginBlockEnd: 128,
-    left: {
-      width: 418,
-    },
-    right: {
-      width: 433,
-    },
-  },
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function FlexSectionTestimonials({testimonials, className}: any) {
-  const {colorMode} = useTheme()
-
-  const isMinimalVariant = testimonials.fields.variant === 'minimal'
-
-  const set =
-    !isMinimalVariant && testimonials.fields.backgroundImageVariant
-      ? backgroundStylesMap[testimonials.fields.backgroundImageVariant]
-      : undefined
-
-  return set ? (
-    <Grid>
-      <Grid.Column>
-        <Box
-          className={clsx(styles['position-relative'], set.className)}
-          marginBlockStart={{narrow: 'none', wide: set.marginBlockStart}}
-          marginBlockEnd={{narrow: 'none', wide: set.marginBlockEnd}}
-        >
-          <Image
-            src={colorMode === 'light' ? startShapeLight : startShapeDark}
-            alt=""
-            className={clsx(styles.testimonialBackgroundImageShape, styles.left)}
-            width={set.left.width}
-            loading="lazy"
-          />
-          <Image
-            src={colorMode === 'light' ? endShapeLight : endShapeDark}
-            alt=""
-            className={clsx(styles.testimonialBackgroundImageShape, styles.right)}
-            width={set.right.width}
-            loading="lazy"
-          />
-
-          <ContentfulTestimonials className={className} testimonials={testimonials} />
-        </Box>
-      </Grid.Column>
-    </Grid>
-  ) : (
-    <ContentfulTestimonials className={className} testimonials={testimonials} />
-  )
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function ContentfulTestimonials({testimonials, className}: any) {
-  const isSingleTestimonial = testimonials.fields.testimonialCount === 1
-  // The default variant is 'minimal', so if the variant is not set, we should treat it as 'minimal'.
-  const isMinimalVariant = testimonials.fields.variant === 'minimal'
-  const span = isSingleTestimonial ? (isMinimalVariant ? 10 : 12) : 6
-
-  return (
-    <Grid className={className}>
-      {Array.from({length: testimonials.fields.testimonialCount}).map((_, i) => (
-        <Grid.Column
-          start={isSingleTestimonial && isMinimalVariant ? {medium: 2} : undefined}
-          span={{medium: span}}
-          key={i}
-        >
-          {testimonials.fields.variant === 'frosted-glass' ? (
-            <FrostedGlassVFX className={styles['height-full']}>
-              <ContentfulTestimonialContent
-                variant="default"
-                size={isSingleTestimonial ? 'large' : 'small'}
-                quoteMarkColor={testimonials.fields.quoteMarkColor}
-                displayedAuthorImage={testimonials.fields.displayedAuthorImage}
-              />
-            </FrostedGlassVFX>
-          ) : (
-            <ContentfulTestimonialContent
-              className={styles['height-full']}
-              size={isSingleTestimonial ? 'large' : 'small'}
-              quoteMarkColor={testimonials.fields.quoteMarkColor}
-              displayedAuthorImage={testimonials.fields.displayedAuthorImage}
-            />
-          )}
-        </Grid.Column>
-      ))}
-    </Grid>
-  )
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const ContentfulTestimonialContent = ({className, size, variant, quoteMarkColor, displayedAuthorImage}: any) => {
-  return (
-    <Testimonial className={className} size={size} variant={variant} quoteMarkColor={quoteMarkColor}>
-      <Testimonial.Quote>
-        <em>GitHub helps us ensure</em> that we have our security controls baked into our pipelines all the way from the
-        first line of code we&apos;re writing.
-      </Testimonial.Quote>
-
-      <Testimonial.Name position="Staff Security Engineer">David Ross</Testimonial.Name>
-
-      {displayedAuthorImage === 'logo' ? (
-        <Testimonial.Logo>
-          <img
-            src="https://github.githubassets.com/images/modules/logos_page/GitHub-Logo.png"
-            alt="GitHub logo"
-            width={60}
-          />
-        </Testimonial.Logo>
-      ) : displayedAuthorImage === 'avatar' ? (
-        <Testimonial.Avatar alt="The testimonial author" src={monaAvatar} />
-      ) : /**
-       * If neither logo nor author's photo exist or none is selected, we don't render anything.
-       */
-      null}
-    </Testimonial>
-  )
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function ContentfulRiver({rivers, className}: any) {
-  /**
-   * We use an empty fragment if `props.content.cta` is not defined for compliance
-   * with `River.Content` types (`River.Content` does not accept `null` as children).
-   */
-  const ctaLink = rivers.fields.hasCta ? (
-    <Link variant={rivers.fields.ctaVariant ?? 'accent'} href="#">
-      Learn more
-    </Link>
-  ) : (
-    <></>
-  )
-
-  const getTrailingComponent = rivers.fields.hasTrailingComponent
-    ? () => (
-        <Timeline>
-          <Timeline.Item>
-            GitHub Codespaces offers a complete dev environment in seconds, so you can code, build, test, and open pull
-            requests from any repo anywhere.
-          </Timeline.Item>
-          <Timeline.Item>
-            GitHub Copilot is your AI pair programmer that empowers you to complete tasks 55% faster by turning natural
-            language prompts into coding suggestions.
-          </Timeline.Item>
-          <Timeline.Item>
-            GitHub Actions automates your build, test, and deployment workflow with simple and secure CI/CD.
-          </Timeline.Item>
-        </Timeline>
-      )
-    : undefined
-
-  return (
-    <River className={className} align={rivers.fields.align} imageTextRatio={rivers.fields.imageTextRatio}>
-      <River.Visual hasShadow={rivers.fields.hasShadow} className={styles['width-full']}>
-        {rivers.fields.visualType === 'image' ? (
-          <img src={placeholderImage} alt="placeholder, blank area with a gray background color" />
-        ) : rivers.fields.visualType === 'video' ? (
-          <video
-            loop
-            playsInline
-            autoPlay
-            muted
-            poster="https://github.githubassets.com/images/modules/site/issues/issue-tasks-progress-placeholder.png"
-          >
-            <source
-              type="video/mp4; codecs=hevc,mp4a.40.2"
-              src="https://github.githubassets.com/images/modules/site/issues/issue-tasks-progress.hevc.mp4"
-            />
-            <source
-              type="video/mp4; codecs=avc1.4D401E,mp4a.40.2"
-              src="https://github.githubassets.com/images/modules/site/issues/issue-tasks-progress.h264.mp4"
-            />
-          </video>
-        ) : null}
-      </River.Visual>
-
-      <River.Content trailingComponent={getTrailingComponent}>
-        {rivers.fields.hasLabel && (
-          <Label
-            size={rivers.fields.labelSize}
-            color={rivers.fields.labelColor}
-            {...(rivers.fields.hasLeadingVisual ? {leadingVisual: <ZapIcon />} : {})}
-          >
-            Label
-          </Label>
-        )}
-
-        <Heading>Heading</Heading>
-        <Text>
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit. In sapien sit ullamcorper id. Aliquam luctus
-          <sup>
-            <InlineLink
-              href=""
-              className={clsx(styles.footnoteInlineLink, styles.footnoteSizeSmall)}
-              aria-label="Footnote 5"
-            >
-              5
-            </InlineLink>
-          </sup>{' '}
-          sed turpis felis nam pulvinar risus elementum.
-        </Text>
-
-        {ctaLink}
-      </River.Content>
-    </River>
-  )
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function ContentfulRiverBreakout({rivers, className}: any) {
-  /**
-   * We use an empty fragment if `props.content.cta` is not defined for compliance
-   * with `River.Content` types (`River.Content` does not accept `null` as children).
-   */
-  const ctaLink = rivers.fields.hasCta ? (
-    <Link variant={rivers.fields.ctaVariant ?? 'accent'} href="#">
-      Learn more
-    </Link>
-  ) : (
-    <></>
-  )
-
-  const getTrailingComponent = rivers.fields.hasTrailingComponent
-    ? () => (
-        <Timeline>
-          <Timeline.Item>
-            GitHub Codespaces offers a complete dev environment in seconds, so you can code, build, test, and open pull
-            requests from any repo anywhere.
-          </Timeline.Item>
-          <Timeline.Item>
-            GitHub Copilot is your AI pair programmer that empowers you to complete tasks 55% faster by turning natural
-            language prompts into coding suggestions.
-          </Timeline.Item>
-          <Timeline.Item>
-            GitHub Actions automates your build, test, and deployment workflow with simple and secure CI/CD.
-          </Timeline.Item>
-        </Timeline>
-      )
-    : undefined
-
-  return (
-    <RiverBreakout className={className}>
-      <RiverBreakout.A11yHeading>Accessible heading</RiverBreakout.A11yHeading>
-      <RiverBreakout.Visual hasShadow={rivers.fields.riverHasShadow}>
-        {rivers.fields.visualType === 'image' ? (
-          <img src={placeholderImage} alt="placeholder, blank area with a gray background color" />
-        ) : rivers.fields.visualType === 'video' ? (
-          <video
-            loop
-            playsInline
-            autoPlay
-            muted
-            poster="https://github.githubassets.com/images/modules/site/issues/issue-tasks-progress-placeholder.png"
-          >
-            <source
-              type="video/mp4; codecs=hevc,mp4a.40.2"
-              src="https://github.githubassets.com/images/modules/site/issues/issue-tasks-progress.hevc.mp4"
-            />
-            <source
-              type="video/mp4; codecs=avc1.4D401E,mp4a.40.2"
-              src="https://github.githubassets.com/images/modules/site/issues/issue-tasks-progress.h264.mp4"
-            />
-          </video>
-        ) : null}
-      </RiverBreakout.Visual>
-      <RiverBreakout.Content trailingComponent={getTrailingComponent}>
-        <Text>
-          <em>Accelerate your workflows</em> and scale your business fast with access to millions of open source
-          projects on GitHub, the largest
-          <sup>
-            <InlineLink
-              href=""
-              className={clsx(styles.footnoteInlineLink, styles.footnoteSizeLarge)}
-              aria-label="Footnote 6"
-            >
-              6
-            </InlineLink>
-          </sup>{' '}
-          source code host.
-        </Text>
-
-        {ctaLink}
-      </RiverBreakout.Content>
-    </RiverBreakout>
-  )
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function ContentfulRiverAccordion({rivers, className}: any) {
-  return (
-    <RiverAccordion align={rivers.fields.align} className={className}>
-      {Array.from({length: 3}).map((_, i) => {
-        return (
-          <RiverAccordion.Item key={i}>
-            <RiverAccordion.Heading>Heading {i + 1}</RiverAccordion.Heading>
-            <RiverAccordion.Content>
-              <Text>
-                Lorem ipsum dolor sit amet consectetur adipisicing elit. Rerum repellendus veniam
-                <sup>
-                  <InlineLink
-                    href=""
-                    className={clsx(styles.footnoteInlineLink, styles.footnoteSizeSmall)}
-                    aria-label="Footnote 7"
-                  >
-                    7
-                  </InlineLink>
-                </sup>{' '}
-                repellat unde ex aut minus iusto, ad a vero optio at eligendi, ratione commodi minima! Laboriosam quas
-                numquam totam.
-              </Text>
-              {rivers.fields.hasCta ? (
-                <Link variant={rivers.fields.ctaVariant ?? 'accent'} href="#">
-                  Learn more
-                </Link>
-              ) : null}
-            </RiverAccordion.Content>
-            <RiverAccordion.Visual className={styles['width-full']}>
-              <img
-                src={placeholderImage}
-                alt="placeholder, blank area with a gray background color"
-                className={styles['width-full']}
-              />
-            </RiverAccordion.Visual>
-          </RiverAccordion.Item>
-        )
-      })}
-    </RiverAccordion>
   )
 }
