+++ packages/react/src/recipes/FlexTemplate/FlexSection/components/index.ts
@@ -1,0 +1,6 @@
+export * from './FlexSectionTestimonials'
+export * from './ContentfulTestimonials'
+export * from './ContentfulTestimonialContent'
+export * from './ContentfulRiver'
+export * from './ContentfulRiverBreakout'
+export * from './ContentfulRiverAccordion'
